Run kakship asynchronously
When working in a large project like Chromium, starship git modules
like git_status can take a long time (I've experienced up to a second
of delay). Previously, the kakship call was run synchronously, which
blocks the editor while the command is running and prevents user
input. This change makes the kakship call async so movement is still
smooth while the prompt is being created.